backend/requirements.txt
certifi==2024.7.4
chardet==4.0.0
click==8.1.7
Flask==2.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
PyJWT==2.1.0
# pyspark==3.5.0
python-dotenv==0.19.0
requests==2.31
SQLAlchemy==1.4.22
urllib3==1.26.19
Werkzeug==2.2.2
blinker==1.8.2
colorama==0.4.6
neo4j
pytz==2024.1
Flask_Migrate==4.0.7
neomodel==5.3.3
flask_cors